refactor(ai): consolidate action-context modules into single file
Merge action-names.ts, actions-to-client-tools.ts, and tools-to-definitions.ts
into action-context.ts. These modules are purely internal implementation details
of createActionContext—the only public export consumers actually use.

Reorganize file structure:
- Public API (ActionLabel, createActionContext) at top
- Internal helpers (actionsToClientTools, toDefinitions, normalizeSchema) below
- Simplify barrel export to just the 2 public symbols

Reduces file overhead while maintaining clear separation of concerns via
section comments. No functional changes.
