+++ packages/ai/src/index.ts
@@ -1,3 +1,1 @@
 export { type ActionLabel, createActionContext } from './action-context';
-export type { ActionNames } from './action-names';
-export type { ServerToolDefinition } from './tools-to-definitions';
